move wrappedClient to client.go
