+++ service/controller/resource/namespace/client.go
@@ -1,0 +1,23 @@
+package namespace
+
+import (
+	corev1 "k8s.io/api/core/v1"
+	v1 "k8s.io/client-go/deprecated/typed/core/v1"
+)
+
+type wrappedClient struct {
+	client v1.NamespaceInterface
+}
+
+func (c wrappedClient) Create(o metav1.Object) (metav1.Object, error) {
+	return c.client.Create(o.(*corev1.Namespace))
+}
+func (c wrappedClient) Update(o metav1.Object) (metav1.Object, error) {
+	return c.client.Update(o.(*corev1.Namespace))
+}
+func (c wrappedClient) Get(name string, options metav1.GetOptions) (metav1.Object, error) {
+	return c.client.Get(name, options)
+}
+func (c wrappedClient) Delete(name string, options *metav1.DeleteOptions) error {
+	return c.client.Delete(name, options)
+}
