+++ service/controller/resource/namespace/resource.go
@@ -6,7 +6,6 @@ import (
 	"github.com/giantswarm/micrologger"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	v1 "k8s.io/client-go/kubernetes/typed/core/v1"
 
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/generic"
 	"github.com/giantswarm/prometheus-meta-operator/service/key"
@@ -19,23 +18,6 @@ const (
 type Config struct {
 	K8sClient k8sclient.Interface
 	Logger    micrologger.Logger
-}
-
-type wrappedClient struct {
-	client v1.NamespaceInterface
-}
-
-func (c wrappedClient) Create(o metav1.Object) (metav1.Object, error) {
-	return c.client.Create(o.(*corev1.Namespace))
-}
-func (c wrappedClient) Update(o metav1.Object) (metav1.Object, error) {
-	return c.client.Update(o.(*corev1.Namespace))
-}
-func (c wrappedClient) Get(name string, options metav1.GetOptions) (metav1.Object, error) {
-	return c.client.Get(name, options)
-}
-func (c wrappedClient) Delete(name string, options *metav1.DeleteOptions) error {
-	return c.client.Delete(name, options)
 }
 
 func New(config Config) (*generic.Resource, error) {
